product uses mybatis mapper and service

+++ src/main/java/org/abondar/experimental/shoppingcart/product/ProductListResponse.java
@@ -6,6 +6,6 @@ public record ProductListResponse(
         List<ProductResponse> products,
         int limit,
         int offset,
-        int total
+        long total
 ) {
 }
